Added patches that need to load last every time.
Added a directory that holds two files that includes values that I want
applied every time, regardless of what has been edited during the load
up to that point. This is done by being the folder that comes in last
alphabetically, and the patches in include the Module Manager flag
":FINAL". Since the folder is loaded last and the patches are set to
apply last after anything else, these patches will always be the values
used in-game.